Modificaciones de endpoints versionados, renombre de rutas y recursos relacionados
