+++ src/app.js
@@ -3,10 +3,10 @@ import express from 'express'
 import morgan from 'morgan';
 import cors from 'cors'
 import { dbConnect } from './database.js'
-import { routerAffiliations } from './routes/affiliations.routes.js'
-import { routerStates } from './routes/states.routes.js'
-import { routerMunicipalities } from './routes/municipalities.routes.js'
-import { routerColonies } from './routes/colonies.routes.js'
+import { routerAffiliation } from './routes/v1/affiliation.routes.js'
+import { routerState } from './routes/v1/state.routes.js'
+import { routerEntity } from './routes/v1/entity.routes.js'
+import { routerTown } from './routes/v1/town.routes.js'
 
 const PORT = process.env.PORT || 3000
 
@@ -14,10 +14,10 @@ const app = express();
 app.use(cors())
 app.use(morgan('dev'));
 app.use(express.json());
-app.use(routerAffiliations)
-app.use(routerStates)
-app.use(routerMunicipalities)
-app.use(routerColonies)
+app.use('/api/v1', routerAffiliation)
+app.use('/api/v1', routerState)
+app.use('/api/v1', routerEntity)
+app.use('/api/v1', routerTown)
 
 dbConnect().then(() => { console.info('Connect to DB') })
