--- src/controllers/colonies.controllers.js
@@ -1,16 +1,0 @@
-import { Colonies } from '../models/colonies.model.js'
-
-const getAll = async (req, res) => {
-    try {
-        const query = await Colonies.find()
-        if (query) {
-            return query
-                ? res.status(200).json({ status: 200, response: "success", data: query })
-                : res.status(404).json({ status: 404, responseError: "not found" })
-        }
-    } catch (error) {
-        return res.status(500).json({ status: 500, responseError: error })
-    }
-}
-
-export { getAll }
