+++ src/models/entity.model.js
@@ -1,0 +1,18 @@
+import { Schema, model } from 'mongoose';
+
+const entitySchena = new Schema({
+    name: {
+        type: String,
+        required: true,
+        trim: true
+    },
+    state: {
+        type: String,
+        required: true
+    }
+}, {
+    versionKey: false,
+    timestamps: true
+});
+
+export const Entity = model('Entities', entitySchena)
