+++ src/models/state.model.js
@@ -1,0 +1,14 @@
+import { Schema, model } from 'mongoose';
+
+const stateSchena = new Schema({
+    name: {
+        type: String,
+        required: true,
+        trim: true
+    }
+}, {
+    versionKey: false,
+    timestamps: true
+});
+
+export const State = model('States', stateSchena)
