--- src/routes/colonies.routes.js
@@ -1,10 +1,0 @@
-import { Router } from "express"
-import { getAll } from '../controllers/colonies.controllers.js'
-
-const routerColonies = Router()
-
-routerColonies
-
-    .get('/colonies', getAll)
-
-export { routerColonies }
