--- src/routes/municipalities.routes.js
@@ -1,10 +1,0 @@
-import { Router } from "express"
-import { getAll } from '../controllers/municipalities.controllers.js'
-
-const routerMunicipalities = Router()
-
-routerMunicipalities
-
-    .get('/municipalities', getAll)
-
-export { routerMunicipalities }
