--- src/routes/states.routes.js
@@ -1,10 +1,0 @@
-import { Router } from "express"
-import { getAll } from '../controllers/states.controllers.js'
-
-const routerStates = Router()
-
-routerStates
-
-    .get('/states', getAll)
-
-export { routerStates }
