+++ src/routes/v1/entity.routes.js
@@ -1,0 +1,10 @@
+import { Router } from "express"
+import { getAll } from '../../controllers/entity.controllers.js'
+
+const routerEntity = Router()
+
+routerEntity
+
+    .get('/entities', getAll)
+
+export { routerEntity }
