+++ src/routes/v1/state.routes.js
@@ -1,0 +1,10 @@
+import { Router } from "express"
+import { getAll } from '../../controllers/state.controllers.js'
+
+const routerState = Router()
+
+routerState
+
+    .get('/states', getAll)
+
+export { routerState }
