+++ src/routes/v1/town.routes.js
@@ -1,0 +1,10 @@
+import { Router } from "express"
+import { getAll } from '../../controllers/town.controllers.js'
+
+const routerTown = Router()
+
+routerTown
+
+    .get('/towns', getAll)
+
+export { routerTown }
